Added Axis Steps:mm menu to support steps:mm mapping for machines other than ToM.  Firmware functions are now calculated based on these settings. Home offsets, Jog now reports in mm.  Supports Axis X,Y,Z and Extruder.

+++ firmware/src/Motherboard/Command.hh
@@ -41,13 +41,13 @@ void pause(bool pause);
 /// \return True if it is disabled, false if it is enabled.
 bool isPaused();
 
-/// \Pause at >= a Z Position provded in mm
+/// \Pause at >= a Z Position provded in steps
 /// 0 cancels pauseAtZPos
-void pauseAtZPos(float zpos);
+void pauseAtZPos(int32_t zpos);
 
 /// Get the current pauseAtZPos position
-/// \return the z position set for pausing, otherwise 0
-float getPauseAtZPos();
+/// \return the z position set for pausing (in steps), otherwise 0
+int32_t getPauseAtZPos();
 
 /// Check the remaining capacity of the command buffer
 /// \return Amount of space left in the buffer, in bytes
